+++ src/routes/users/name.ts
@@ -1,9 +1,10 @@
 import User from "../../models/user";
-import { threadCl, usersCl } from "../../common";
+import { usersCl } from "../../common";
 import { Static, Type } from "@sinclair/typebox";
 import { FastifyInstance, FastifyPluginOptions, FastifyRequest } from "fastify";
 import regex from "../../lib/regex";
 
+/** get username */
 export default function (
     fastify: FastifyInstance,
     _opts: FastifyPluginOptions,
@@ -12,7 +13,7 @@ export default function (
     const paramsSchema = Type.Object({ id: Type.RegEx(regex.integer) });
 
     fastify.get(
-        "/:id",
+        "/:id/name",
         {
             schema: {
                 params: paramsSchema,
@@ -28,25 +29,12 @@ export default function (
 
             const requestedUser = (await usersCl.findOne(
                 { id },
-                {
-                    projection: {
-                        id: 1,
-                        name: 1,
-                        createdAt: 1,
-                        sex: 1,
-                        role: 1,
-                        _id: 0,
-                    },
-                }
+                { projection: { name: 1, _id: 0 } }
             )) as User;
 
             if (!requestedUser) return res.code(404).send({ error: "User not found" });
 
-            const count = await threadCl.countDocuments({
-                "op.id": requestedUser.id,
-            });
-
-            res.send({ ...requestedUser, count });
+            res.send(requestedUser);
         }
     );
     done();